+++ CHANGELOG.md
@@ -16,6 +16,8 @@ this feature should query their public IP at runtime using a service such as
 checkip.amazon.com, or pass containers a floating IP (or DNS name backed by a
 floating IP) and use placement rules to ensure the container is scheduled at
 that floating IP.
+- Remove the `.q` TLD for Kelda hostnames. Kubernetes doesn't support hostnames
+with TLDs.
 
 Javascript API changes:
 - Remove deprecated methods `allowFrom` and `allow`.